adding call to 'confirmation dialog' + refactoring

+++ src/app/components/coupon-profile/coupon-profile.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmationDialog } from './../../service/confirmation-dialog';
 import { Observable } from 'rxjs';
 import { Company } from './../../models/company';
 import { CustomerService } from './../../service/customer';
@@ -27,6 +28,7 @@ export class CouponProfileComponent implements OnInit {
     private generalService: GeneralService,
     private companyService: CompanyService,
     private customerService: CustomerService,
+    private confirmationDialog: ConfirmationDialog
   ) { }
 
   ngOnInit(): void {
@@ -82,8 +84,16 @@ export class CouponProfileComponent implements OnInit {
   }
 
   deleteCoupon(): void {
-    this.companyService.deleteCoupon(this.coupon.id).subscribe(() => {
-      this.router.navigate([this.authentication.getUrl]);
+    this.confirmationDialog.confirm(
+      'Delete coupon alert',
+      'Are you sure you want to delete this coupon?',
+      'Delete'
+    ).then((confirmed: boolean) => {
+      if (confirmed) {
+        this.companyService.deleteCoupon(this.coupon.id).subscribe(() => {
+          this.router.navigate([this.authentication.getUrl]);
+        });
+      }
     });
   }
 
@@ -98,8 +108,16 @@ export class CouponProfileComponent implements OnInit {
   }
 
   cancelPurchase(): void {
-    this.customerService.removePurchasedCoupon(this.coupon.id).subscribe(() => {
-      this.router.navigate([this.authentication.getUrl]);
+    this.confirmationDialog.confirm(
+      'Purchased coupon removal',
+      'Are you sure you want to remove this coupon?',
+      'Remove'
+    ).then((confirmed: boolean) => {
+      if (confirmed) {
+        this.customerService.removePurchasedCoupon(this.coupon.id).subscribe(() => {
+          this.router.navigate([this.authentication.getUrl]);
+        });
+      }
     });
   }
